Refactor Dashboard component into Home view
